transferred the profile button to the settings page.
