--- app/src/main/res/layout/activity_profile_view.xml
@@ -1,115 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ProfileView">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/purple_339"
-        android:orientation="vertical"
-        android:padding="16dp"
-        tools:layout_editor_absoluteX="0dp"
-        tools:layout_editor_absoluteY="0dp">
-
-        <RelativeLayout
-            android:layout_width="389dp"
-            android:layout_height="35dp"
-            android:background="@color/purple_401"
-            tools:ignore="UselessParent">
-
-            <!-- app:icon="@drawable/back_btn" -->
-
-            <ImageButton
-                android:id="@+id/backBtn"
-                style="@style/Widget.AppCompat.Button.Borderless"
-                android:layout_width="100dp"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="-10dp"
-                android:src="@drawable/back_btn"
-                android:textColor="@color/white"
-                android:visibility="visible" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_centerInParent="true"
-                android:textColor="@color/white"
-                android:textSize="18sp"
-                android:textStyle="bold"
-                android:visibility="visible" />
-        </RelativeLayout>
-
-<!--        <Button-->
-<!--            android:id="@+id/backBtn"-->
-<!--            android:layout_width="match_parent"-->
-<!--            android:layout_height="wrap_content"-->
-<!--            android:text="Back" />-->
-
-        <ImageView
-            android:id="@+id/imageView"
-            android:layout_width="105dp"
-            android:layout_height="105dp"
-            tools:srcCompat="@tools:sample/avatars" />
-
-        <EditText
-            android:id="@+id/etName"
-            android:layout_width="160dp"
-            android:layout_height="wrap_content"
-            android:ems="10"
-            android:layout_gravity="right"
-            android:inputType="text"
-            android:textColor="@color/white"
-            android:textSize="20sp"
-            android:hint="Name" />
-
-        <EditText
-            android:id="@+id/etSurname"
-            android:layout_width="160dp"
-            android:layout_height="wrap_content"
-            android:ems="10"
-            android:layout_gravity="right"
-            android:inputType="text"
-            android:textColor="@color/white"
-            android:textSize="20sp"
-            android:hint="Surname"/>
-
-        <TextView
-            android:id="@+id/textView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textSize="30dp"
-            android:hint="Full Name" />
-
-        <Button
-            android:id="@+id/result"
-            android:layout_width="140dp"
-            android:layout_height="wrap_content"
-            android:layout_gravity="right"
-            android:ems="10"
-            android:onClick="btnSave"
-            android:text="Save to file"/>
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rvPinnedBadges"
-            android:layout_width="wrap_content"
-            android:layout_height="326dp"
-            tools:itemCount="3"
-            tools:listitem="@layout/item_badge">
-
-        </androidx.recyclerview.widget.RecyclerView>
-
-        <Button
-            android:id="@+id/btnAdd"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="btnAdd"
-            android:text="Add" />
-
-
-
-    </LinearLayout>
-</androidx.constraintlayout.widget.ConstraintLayout>
